:recycle: 회원관리 타입 디렉토리 정리
