--- src/types/users/login.d.ts
@@ -1,6 +1,0 @@
-interface LoginType extends Pick<UserInfoType, "username" | "pw"> {}
-
-interface FindIdType extends pick<UserInfoType, "email" | "name"> {}
-
-interface FindPwType
-  extends Pick<UserInfoType, "email" | "username" | "name"> {}
