+++ src/types/users/validate-verify.d.ts
@@ -1,4 +1,19 @@
-import { VERIFICATION } from "./enums";
+import { VALIDATION } from "./enums";
+
+// VALIDATION ########################
+
+interface ValidationList {
+  [VALIDATION.CODE]: boolean;
+  [VALIDATION.EMAIL]: boolean;
+  [VALIDATION.NAME]: boolean;
+  [VALIDATION.USER_ID]: boolean;
+  [VALIDATION.USER_PW]: boolean;
+}
+
+interface LoginValidation
+  extends Pick<ValidationList, VALIDATION.USER_ID | VALIDATION.USER_PW> {}
+
+// VERIFICATION ########################
 
 interface VerificationList {
   [VERIFICATION.CODE]: boolean;
